Backend/SmartPlatform/requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
uuid==1.30
django-enumchoicefield==3.0.1
django-cors-headers==3.13.0
channels==4.0.0
channels-redis==4.0.0
celery==5.2.7
redis==4.5.1
eventlet==0.33.3